pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tts-tester"
version = "0.1.0"
description = "Local Google Cloud Text-to-Speech tester – CLI + optional Streamlit UI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "You"}]

dependencies = [
    "google-cloud-texttospeech>=2.16,<3",
    "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
ui = ["streamlit>=1.30,<2"]

[project.scripts]
tts-tester = "tts_tester.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
